scripts/requirements.txt
requests
PyYAML
certifi